Change maps authentication setup
Changed the ARM template back to just using the primary key for the Map resource authentication. While the better way would have been via AAD, owner permissions are required to create an App Registration and no everyone who may want to try this will have that permission on their subscription.

So while using the primary key is not for a proper solution, it servers the purpose for this sample. Also, it will be retrieved from the KeyVault etc.