Add map to the web app
Updated the web app so that it displays the map. Currently it is set to the centre of the UK roughly.

While the primary key of the Map resource is visible, it is retrieved from the KeyVault as an example on how an App Service can get config from a KeyVault via Managed Identities.